Add files via upload
The project activity involves the setup of a biogas digester for the molasses based distillery with a production capacity of 45 klpd that were both commissioned in 2019. The project activity is also the anaerobic treatment of wastewater for methane recovery. Since the project activity includes generation of steam and power with utilization of recovered biogas (methane), this avoids grid electricity imports thus reducing emissions from thermal sources connected to the project electricity grid system.